# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Ludus Engine API",
    "description" : "Backend engine for 2D games. Content is authored in the editor and served to game clients over the public routes, which are cacheable and validated by ETag.",
    "license" : {
      "name" : "AGPL-3.0-or-later",
      "url" : "https://www.gnu.org/licenses/agpl-3.0.txt"
    },
    "version" : "${project.version}"
  },
  "paths" : {
...
    "/api/v1/admin/api-keys" : {
      "get" : {
        "tags" : [ "API keys" ],
        "summary" : "List keys, including revoked ones",
        "operationId" : "listApiKeys",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Summary"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "API keys" ],
        "summary" : "Mint a key. The response is the only time it is shown.",
        "operationId" : "createApiKey",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreatedResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/audio/{id}" : {
      "get" : {
        "tags" : [ "Public content" ],
        "summary" : "Stream an audio clip",
        "description" : "Cached immutably: a clip never changes under its id, so a client that has it never asks again.",
        "operationId" : "streamAudioClip",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StreamingResponseBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/audio/{id}" : {
      "delete" : {
        "tags" : [ "Audio" ],
        "summary" : "Delete a clip and its bytes",
        "operationId" : "deleteAudioClip",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/admin/api-keys/{id}" : {
      "delete" : {
        "tags" : [ "API keys" ],
        "summary" : "Revoke a key",
        "description" : "The row is kept and stamped, not deleted: when a key stopped working is a question people ask after an incident.",
        "operationId" : "revokeApiKey",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Summary" : {
        "type" : "object",
        "description" : "A wave's indexed fields. The document itself is fetched separately.",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "order" : {
            "type" : "integer",
            "format" : "int32"
          },
          "schemaVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "published" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CreateRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "So a human can tell keys apart",
            "example" : "android-client",
            "maxLength" : 120,
            "minLength" : 0
          }
        },
        "required" : [ "name" ]
      },
      "CreatedResponse" : {
        "type" : "object",
        "description" : "The only response that ever contains the key itself. It is stored as a digest, so it cannot be shown again by anyone, including whoever runs the database.",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "prefix" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "key" : {
            "type" : "string"
          }
        }
      },
...
      "StreamingResponseBody" : { },
...
    }
  }
}